Many changes
It seems that this code was submitted in midst of a migration of template
engines, so I did my best to fit the old code to the newer templating scheme.
There were also some parts that felt unimplemented, so I took the liberty of
implementing them, like the user management screen. I've also added a user group
management screen that doubles as permission group. If you're on the admin group
you'll be allowed to see admin-only plugins, else you'll only see non-admin stuff.

I've also fixed the pagination tool on the CDR plugin (which didn't respected
previous search results) and added a date search field.

You can also find the schema for the database on root/database.sql.