click-2-dial dialog follows the call progress
[Plugins] : click-2-dial waits until the extension has
picked up to confirm if the originate call was successful.
In case of multiple channels, it will also hangup others
channels as soon as one picks up. This functionnality
requires javascript. If it is disabled, the click-2-dial
function sends the originate request for all channel but
is unable to follow the call progress.