Fixed : Wrong referrer for add/edit dialogs.
When submitting a form inside an add/edit dialog, the browser
wes redirected back to the same page. This is because the
referrer parameter is set from the current url. With the
keep-referrer option enabled, the original referrer url is
preserved and the browser is then redirected to the previous
page.

* [Template engine] : Added keep-referrer option for form tags.